problem-generator-task-package/rag/requirements.txt
pdfplumber
pandas
tqdm
faiss-cpu
sentence-transformers
rank_bm25
numpy
